create component AuthLayout.jsx
Authentication Layout(AuthLayout.jsx) is a mechanism that helps protect routes/pages. Its a protected container for values, and decides whether to show the values or not.

We can change the name of function, like we changed name from AuthLayout() into Protected(). We're giving name Protected() because we will render conditionally and check whether to render its children or not.

Signed-off-by: TBS96 <9tbs6@proton.me>